another timespan

+++ AspNet/MetricsManager/MetricsAgent/DAL/DataTimeOffsetHandler.cs
@@ -10,7 +10,7 @@ namespace MetricsAgent.DAL
     public class DataTimeOffsetHandler : SqlMapper.TypeHandler<DateTimeOffset>
     {
         public override DateTimeOffset Parse(object value)
-        => DateTimeOffset.FromUnixTimeSeconds((Int32)value);
+        => DateTimeOffset.FromUnixTimeSeconds((long)value);
 
         public override void SetValue(IDbDataParameter parameter, DateTimeOffset value)
         => parameter.Value = value;
