+++ AspNet/MetricsManager/MetricsAgent/DAL/TimeSpanHandler.cs
@@ -7,7 +7,7 @@ namespace MetricsAgent.DAL
     public class TimeSpanHandler : SqlMapper.TypeHandler<TimeSpan>
     {
         public override TimeSpan Parse(object value)
-            => TimeSpan.FromSeconds((Int32)value);
+            => TimeSpan.FromSeconds((long)value);
 
         public override void SetValue(IDbDataParameter parameter, TimeSpan value)
             => parameter.Value = value;
